+++ analyzers/src/SonarAnalyzer.CSharp/SymbolicExecution/Sonar/Constraints/InitializationVectorConstraint.cs
@@ -18,15 +18,18 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarAnalyzer.SymbolicExecution.Constraints;
+using SonarAnalyzer.SymbolicExecution.Constraints;
 
-internal class ByteArrayConstraint : SymbolicConstraint
+namespace SonarAnalyzer.SymbolicExecution.Sonar.Constraints
 {
-    public static readonly ByteArrayConstraint Constant = new(ConstraintKind.ByteArrayConstant);
-    public static readonly ByteArrayConstraint Modified = new(ConstraintKind.ByteArrayModified);
+    internal class InitializationVectorConstraint : SymbolicConstraint
+    {
+        public static readonly InitializationVectorConstraint NotInitialized = new(ConstraintKind.InitializationVectorNotInitialized);
+        public static readonly InitializationVectorConstraint Initialized = new(ConstraintKind.InitializationVectorInitialized);
 
-    public override SymbolicConstraint Opposite =>
-        this == Modified ? Constant : Modified;
+        public override SymbolicConstraint Opposite =>
+            this == Initialized ? NotInitialized : Initialized;
 
-    private ByteArrayConstraint(ConstraintKind kind) : base(kind) { }
+        private InitializationVectorConstraint(ConstraintKind kind) : base(kind) { }
+    }
 }
