+++ analyzers/tests/SonarAnalyzer.UnitTest/SymbolicExecution/Sonar/Constraints/InitializationVectorConstraintTest.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using SonarAnalyzer.SymbolicExecution.Constraints;
+using SonarAnalyzer.SymbolicExecution.Sonar.Constraints;
 
 namespace SonarAnalyzer.UnitTest.SymbolicExecution.Sonar.Constraints
 {
